refactor(web): route org seat updates through Core API
Call coreClient.updateOrganizationSubscriptionSeats from the subscription
action directly and drop the service passthrough so web auth Prisma
strangler can shed one more write path before cutover.
