+++ apps/web/src/lib/services/__tests__/organization-subscription.service.test.ts
@@ -22,27 +22,6 @@ const resolveActiveSubscriptionByReferenceIdMock = vi.fn();
 const resolveOrganizationBillingPlanMock = vi.fn();
 const prismaTransactionMock = vi.fn();
 const updateSubscriptionRecordMock = vi.fn();
-const updateOrganizationSubscriptionSeatsMock = vi.fn();
-
-class MockCoreApiRequestError extends Error {
-  kind?: string;
-  status?: number;
-
-  constructor(message: string, options?: { kind?: string; status?: number }) {
-    super(message);
-    this.name = "CoreApiRequestError";
-    this.kind = options?.kind;
-    this.status = options?.status;
-  }
-}
-
-vi.mock("@/lib/clients/core.client", () => ({
-  CoreApiRequestError: MockCoreApiRequestError,
-  coreClient: {
-    updateOrganizationSubscriptionSeats: (...args: unknown[]) =>
-      updateOrganizationSubscriptionSeatsMock(...args),
-  },
-}));
 
 vi.mock("@sokosumi/database/repositories", () => ({
   memberRepository: {
@@ -203,168 +182,7 @@ describe("organizationSubscriptionService", () => {
       await organizationSubscriptionService.ensureCanAcceptInvitation("org-1");
 
       expect(getAssignedMemberCountMock).not.toHaveBeenCalled();
-      expect(updateOrganizationSubscriptionSeatsMock).not.toHaveBeenCalled();
       expect(updateSubscriptionRecordMock).not.toHaveBeenCalled();
-    });
-  });
-
-  describe("updateOrganizationSeatsImmediately", () => {
-    it("returns the seat count resolved by core", async () => {
-      updateOrganizationSubscriptionSeatsMock.mockResolvedValue({
-        data: { seats: 6 },
-      });
-
-      const { organizationSubscriptionService } = await import(
-        "../organization-subscription.service"
-      );
-
-      await expect(
-        organizationSubscriptionService.updateOrganizationSeatsImmediately(
-          "user-1",
-          "org-1",
-          6,
-        ),
-      ).resolves.toEqual({
-        seats: 6,
-      });
-
-      expect(updateOrganizationSubscriptionSeatsMock).toHaveBeenCalledWith(
-        "org-1",
-        6,
-      );
-    });
-
-    it("maps a core 403 to the owner/admin FORBIDDEN error", async () => {
-      updateOrganizationSubscriptionSeatsMock.mockRejectedValue(
-        new MockCoreApiRequestError("You must be owner, admin", {
-          status: 403,
-        }),
-      );
-
-      const { organizationSubscriptionService } = await import(
-        "../organization-subscription.service"
-      );
-
-      await expect(
-        organizationSubscriptionService.updateOrganizationSeatsImmediately(
-          "user-1",
-          "org-1",
-          3,
-        ),
-      ).rejects.toThrow(
-        "Only organization owners and admins can manage subscriptions",
-      );
-    });
-
-    it("maps a missing organization to the owner/admin FORBIDDEN error", async () => {
-      updateOrganizationSubscriptionSeatsMock.mockRejectedValue(
-        new MockCoreApiRequestError("Organization not found", {
-          status: 404,
-        }),
-      );
-
-      const { organizationSubscriptionService } = await import(
-        "../organization-subscription.service"
-      );
-
-      await expect(
-        organizationSubscriptionService.updateOrganizationSeatsImmediately(
-          "user-1",
-          "org-1",
-          3,
-        ),
-      ).rejects.toThrow(
-        "Only organization owners and admins can manage subscriptions",
-      );
-    });
-
-    it("maps the organization_not_found kind to FORBIDDEN even when the message is reworded", async () => {
-      updateOrganizationSubscriptionSeatsMock.mockRejectedValue(
-        new MockCoreApiRequestError("We could not find that organization", {
-          kind: "organization_not_found",
-          status: 404,
-        }),
-      );
-
-      const { organizationSubscriptionService } = await import(
-        "../organization-subscription.service"
-      );
-
-      await expect(
-        organizationSubscriptionService.updateOrganizationSeatsImmediately(
-          "user-1",
-          "org-1",
-          3,
-        ),
-      ).rejects.toThrow(
-        "Only organization owners and admins can manage subscriptions",
-      );
-    });
-
-    it("keeps core's message when no active subscription exists", async () => {
-      updateOrganizationSubscriptionSeatsMock.mockRejectedValue(
-        new MockCoreApiRequestError(
-          "An active organization subscription is required before updating seats.",
-          { status: 400 },
-        ),
-      );
-
-      const { organizationSubscriptionService } = await import(
-        "../organization-subscription.service"
-      );
-
-      await expect(
-        organizationSubscriptionService.updateOrganizationSeatsImmediately(
-          "user-1",
-          "org-1",
-          3,
-        ),
-      ).rejects.toThrow(
-        "An active organization subscription is required before updating seats.",
-      );
-    });
-
-    it("keeps core's message when decreasing seats below assigned count", async () => {
-      updateOrganizationSubscriptionSeatsMock.mockRejectedValue(
-        new MockCoreApiRequestError(
-          "Purchased seats (3) must be at least 4 to cover all assigned members",
-          { status: 400 },
-        ),
-      );
-
-      const { organizationSubscriptionService } = await import(
-        "../organization-subscription.service"
-      );
-
-      await expect(
-        organizationSubscriptionService.updateOrganizationSeatsImmediately(
-          "user-1",
-          "org-1",
-          3,
-        ),
-      ).rejects.toThrow(
-        "Purchased seats (3) must be at least 4 to cover all assigned members",
-      );
-    });
-
-    it("rethrows unexpected core errors unchanged", async () => {
-      const coreError = new MockCoreApiRequestError(
-        "Unable to update organization subscription seats: missing Stripe subscription item",
-        { status: 500 },
-      );
-      updateOrganizationSubscriptionSeatsMock.mockRejectedValue(coreError);
-
-      const { organizationSubscriptionService } = await import(
-        "../organization-subscription.service"
-      );
-
-      await expect(
-        organizationSubscriptionService.updateOrganizationSeatsImmediately(
-          "user-1",
-          "org-1",
-          3,
-        ),
-      ).rejects.toBe(coreError);
     });
   });
